Finish removing scss
