+++ rollup.config.js
@@ -1,18 +1,11 @@
-import autoprefixer from 'autoprefixer';
 import babel from 'rollup-plugin-babel';
 import browsersync from 'rollup-plugin-browsersync';
 import commonjs from 'rollup-plugin-commonjs';
-import css from 'rollup-plugin-css-only';
-import cssnano from 'cssnano';
 import { eslint } from 'rollup-plugin-eslint';
 import filesize from 'rollup-plugin-filesize';
-import fs from 'fs';
 import license from 'rollup-plugin-license';
-import postcss from 'postcss';
 import replace from 'rollup-plugin-replace';
 import resolve from 'rollup-plugin-node-resolve';
-import sass from 'rollup-plugin-sass';
-import stylelint from 'rollup-plugin-stylelint';
 import { terser } from 'rollup-plugin-terser';
 import visualizer from 'rollup-plugin-visualizer';
 
@@ -21,39 +14,9 @@ const banner = ['/*!', pkg.name, pkg.version, '*/\n'].join(' ');
 
 const env = process.env.DEVELOPMENT ? 'development' : 'production';
 
-const sassOptions = {
-  output(styles, styleNodes) {
-    fs.mkdirSync('dist/css', { recursive: true }, (err) => {
-      if (err) {
-        throw err;
-      }
-    });
-
-    styleNodes.forEach(({ id, content }) => {
-      const scssName = id.substring(id.lastIndexOf('/') + 1, id.length);
-      const name = scssName.split('.')[0];
-      fs.writeFileSync(`dist/css/${name}.css`, content);
-    });
-  },
-  processor: css => postcss([
-    autoprefixer({
-      grid: false
-    }),
-    cssnano()
-  ])
-    .process(css)
-    .then(result => result.css)
-};
-
 const plugins = [
   resolve(),
   commonjs(),
-  stylelint({
-    fix: false,
-    include: ['src/**.scss'],
-    syntax: 'scss',
-    quiet: false
-  }),
   eslint(),
   babel({
     exclude: 'node_modules/**',
@@ -61,22 +24,11 @@ const plugins = [
   }),
   replace({
     'process.env.NODE_ENV': JSON.stringify(env)
-  }),
-  css({ output: false })
+  })
 ];
-
-if (!process.env.DEVELOPMENT) {
-  plugins.push(
-    sass({
-      output: false
-    })
-  );
-}
 
 // If we are running with --environment DEVELOPMENT, serve via browsersync for local development
 if (process.env.DEVELOPMENT) {
-  plugins.push(sass(sassOptions));
-
   plugins.push(
     browsersync({
       host: 'localhost',
@@ -151,8 +103,6 @@ if (!process.env.DEVELOPMENT) {
           exclude: 'node_modules/**',
           runtimeHelpers: true
         }),
-        sass(sassOptions),
-        css({ output: false }),
         replace({
           'process.env.NODE_ENV': JSON.stringify(env)
         }),
