+++ src/js/shepherd.js
@@ -1,5 +1,3 @@
-import '../scss/shepherd-theme-default.scss';
-
 import { Evented } from './evented.js';
 import { Step } from './step.js';
 import { Shepherd, Tour } from './tour.js';
