--- src/scss/_base.scss
@@ -1,35 +1,0 @@
-@import 'variables';
-@import 'modal-overlay';
-
-.shepherd-element {
-  .shepherd-content {
-    .shepherd-title {
-      color: map_get($shepherd-theme-text-colors, header);
-    }
-
-    .shepherd-cancel-link {
-      color: lighten(map_get($shepherd-theme-text-colors, header), 70);
-
-      &:hover {
-        color: map_get($shepherd-theme-text-colors, header);
-      }
-    }
-
-    .shepherd-text {
-      color: map_get($shepherd-theme-text-colors, text);
-
-      a,
-      a:visited,
-      a:active {
-        border-bottom: 1px dotted;
-        border-bottom-color: map_get($shepherd-theme-text-colors, text);
-        color: map_get($shepherd-theme-text-colors, text);
-        text-decoration: none;
-
-        &:hover {
-          border-bottom-style: solid;
-        }
-      }
-    }
-  }
-}
