--- src/scss/_modal-overlay.scss
@@ -1,4 +1,0 @@
-.shepherd-modal.shepherd-enabled {
-  position: relative;
-  z-index: 9998;
-}
