--- src/scss/_variables.scss
@@ -1,21 +1,0 @@
-@import 'functions';
-
-// Theme colors
-$shepherd-theme-primary: #3288e6 !default;
-$shepherd-theme-secondary: desaturate(lighten($shepherd-theme-primary, 40), 70) !default;
-$shepherd-text-background: #ffffff !default;
-$shepherd-header-background: darken($shepherd-text-background, 10) !default;
-
-$shepherd-theme-text-colors: (
-  primary: rgba(set-text-color($shepherd-theme-primary), 0.75),
-  secondary: rgba(set-text-color($shepherd-theme-secondary), 0.75),
-  header: rgba(set-text-color($shepherd-header-background), 0.75),
-  text: rgba(set-text-color($shepherd-text-background), 0.75)
-) !default;
-
-// Theme options
-$shepherd-element-z-index: 9999 !default;
-
-// Popper options
-$arrow-pointer-events: none !default;
-$arrow-size: 2.1 !default;
