--- src/scss/shepherd-theme-default.scss
@@ -1,1 +1,0 @@
-@import 'base';
